Day 3
I have only done part 1. But part 2 is simmilar and I know what to do,
the only problem is converting a number to base 3. For part 1 I used
the builtin toBinary function, but a toBase3 doesnt exist, so I will
need to make one myself. But I cannot be bothered right now, once the
to toBase3 function is done it should be easy to finnish it. There
maybe perfomance issues for part 2 though, as part 1 took a couple of
seconds and we are adding more operations for part 2.